+++ src/util/oauth.util.ts
@@ -2,8 +2,8 @@ import * as jwt from "jsonwebtoken";
 import * as simpleOauth from "simple-oauth2";
 import { OAuthClient, AccessToken } from "simple-oauth2";
 import * as unixTime from "unix-time";
-import request from "sync-request";
 import * as getPem from "rsa-pem-from-mod-exp";
+import axios from "axios";
 import "@babel/polyfill";
 import { OAuthUser } from "../types";
 
@@ -81,19 +81,29 @@ class OAuthUtil {
    * any public keys that maybe used to sign tokens
    * @method{getOpenIdConfig}
    */
-  getOpenIdConfig() {
+  async getOpenIdConfig() {
     try {
-      const response = request("GET", this.openIdDiscoveryUrl).getBody("utf8");
-      this.openIdConfig = JSON.parse(response);
+      const discoveryClient = axios.create({
+        baseURL: this.openIdDiscoveryUrl,
+        withCredentials: false
+      });
+      const response = await discoveryClient
+            .get("/")
+            .then((res: any) => res.data)
+      this.openIdConfig = response;
       this.jwtAlgorithm = this.openIdConfig.id_token_signing_alg_values_supported;
       this.issuer = this.openIdConfig.issuer;
       this.logOutUrl = this.openIdConfig.end_session_endpoint;
 
       if (this.openIdConfig.jwks_uri) {
-        const keyResponse = request("GET", this.openIdConfig.jwks_uri).getBody(
-          "utf8"
-        );
-        this.keys = JSON.parse(keyResponse).keys;
+        const jwksClient = axios.create({
+          baseURL: this.openIdConfig.jwks_uri,
+          withCredentials: false
+        });
+        const keyResponse = await jwksClient
+            .get("/")
+            .then((res: any) => res.data)
+        this.keys = keyResponse.keys;
       }
     } catch (error) {
       throw error;
@@ -106,10 +116,10 @@ class OAuthUtil {
    * @param oauthConfig
    * @returns o an instance of the OAuthUtil
    */
-  static init(oauthConfig:OAuthConfig) {
+  static async init(oauthConfig:OAuthConfig) {
     try {
       const o = new OAuthUtil(oauthConfig);
-      o.getOpenIdConfig();
+      await o.getOpenIdConfig();
       // get tokenHost
 
       const credentials = {